Move all the composables to shared module and set them internal due to iOS limitations. Change project structure to the jetbrains setup
