--- HelloWorld/src/androidMain/kotlin/com/pablichj/incubator/uistate3/example/helloWorld/MainActivity.kt
@@ -1,23 +1,0 @@
-package com.pablichj.incubator.uistate3.example.helloWorld
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.runtime.CompositionLocalProvider
-import com.pablichj.incubator.uistate3.node.AndroidBackPressDispatcher
-import com.pablichj.incubator.uistate3.node.LocalBackPressedDispatcher
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            CompositionLocalProvider(
-                LocalBackPressedDispatcher provides AndroidBackPressDispatcher(this@MainActivity),
-            ) {
-                ComposeApp(
-                    onExit = { finish() }
-                )
-            }
-        }
-    }
-}
