+++ androidApp/src/androidMain/AndroidManifest.xml
