+++ androidApp/src/androidMain/kotlin/com/pablichj/incubator/uistate3/example/helloWorld/MainActivity.kt
@@ -1,0 +1,14 @@
+package com.pablichj.incubator.uistate3.example.helloWorld
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            HelloWorldApp(this)
+        }
+    }
+}
