+++ desktopApp/build.gradle.kts
@@ -1,0 +1,44 @@
+plugins {
+    kotlin("multiplatform")
+    id("org.jetbrains.compose")
+}
+
+kotlin {
+    jvm {}
+    sourceSets {
+        val jvmMain by getting  {
+            dependencies {
+                implementation(project(":shared"))
+                implementation(compose.desktop.common)
+                implementation(compose.desktop.currentOs)
+                //implementation("org.jetbrains.kotlinx:kotlinx-coroutines-swing:1.6.4")
+            }
+        }
+    }
+}
+
+compose.desktop {
+    application {
+        mainClass = "MainKt"
+    }
+}
+
+/*compose.desktop {
+    application {
+        mainClass = "com.pablichj.incubator.uistate3.example.helloWorld.MainKt"
+
+        nativeDistributions {
+            targetFormats(org.jetbrains.compose.desktop.application.dsl.TargetFormat.Dmg,
+                org.jetbrains.compose.desktop.application.dsl.TargetFormat.Msi,
+                org.jetbrains.compose.desktop.application.dsl.TargetFormat.Deb)
+            packageName = "Hello World"
+            packageVersion = "1.0.0"
+
+            windows {
+                menuGroup = "UiState3 Examples"
+                // see https://wixtoolset.org/documentation/manual/v3/howtos/general/generate_guids.html
+                upgradeUuid = "18159995-d967-4CD2-8885-77BFA97CFA9F"
+            }
+        }
+    }
+}*/
