+++ desktopApp/src/jvmMain/kotlin/Main.kt
@@ -1,0 +1,13 @@
+import androidx.compose.ui.unit.DpSize
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.window.WindowState
+import androidx.compose.ui.window.singleWindowApplication
+import com.pablichj.incubator.uistate3.example.helloWorld.HelloWorldApp
+
+fun main() =
+    singleWindowApplication(
+        title = "Chat",
+        state = WindowState(size = DpSize(500.dp, 800.dp))
+    ) {
+        HelloWorldApp()
+    }
