+++ iosApp/Podfile
@@ -1,4 +1,4 @@
-target 'FallingBalls' do
+target 'HelloWorld' do
   use_frameworks!
   platform :ios, '14.1'
   pod 'shared', :path => '../shared'
