+++ jsApp/build.gradle.kts
@@ -1,0 +1,24 @@
+plugins {
+    kotlin("multiplatform")
+    id("org.jetbrains.compose")
+}
+
+kotlin {
+    js(IR) {
+        browser()
+        binaries.executable()
+    }
+    sourceSets {
+        val jsMain by getting  {
+            dependencies {
+                implementation(project(":shared"))
+                implementation(compose.web.core)
+            }
+        }
+    }
+}
+
+compose.experimental {
+    web.application {}
+}
+
