+++ jsApp/src/jsMain/resources/index.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Hello World</title>
+    <script src="skiko.js"> </script>
+    <link type="text/css" rel="stylesheet" href="styles.css" />
+</head>
+<body>
+    <div>
+        <canvas id="ComposeTarget" width="600" height="600"></canvas>
+    </div>
+    <script src="jsApp.js"> </script>
+</body>
+</html>
