+++ shared/src/commonMain/kotlin/com.pablichj.incubator.uistate3/node/Node.kt
@@ -113,7 +113,7 @@ abstract class Node : Lifecycle {
     // endregion
 
     @Composable
-    abstract fun Content(modifier: Modifier)
+    internal abstract fun Content(modifier: Modifier)
 
     sealed interface LifecycleState {
         object Created : LifecycleState
