+++ shared/src/commonMain/kotlin/com.pablichj.incubator.uistate3/node/topbar/TopBarView.kt
@@ -15,7 +15,7 @@ import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.unit.dp
 
 @Composable
-fun TopBar(
+internal fun TopBar(
     topBarState: ITopBarState
 ) {
